+++ PowerShell/profile.ps1
@@ -9,13 +9,13 @@ $galleryinfo = Get-PSRepository | Where-Object { $_.Name -eq "PSGallery" }
 if (-not($galleryinfo.InstallationPolicy.Equals("Trusted"))) { Set-PSRepository -Name PSGallery -InstallationPolicy Trusted }
 
 # Import Modules
-Import-Module posh-git
-Import-Module oh-my-posh
-Import-Module Terminal-Icons
-Import-Module WslInterop
-Import-Module PSWindowsUpdate
-Import-Module PSWriteColor
-Import-Module WieldingLs
+# Import-Module posh-git
+# Import-Module oh-my-posh
+# Import-Module Terminal-Icons
+# Import-Module WslInterop
+# Import-Module PSWindowsUpdate
+# Import-Module PSWriteColor
+# Import-Module WieldingLs
 
 # Enable Posh-Git
 $env:POSH_GIT_ENABLED = $true